requirements.txt
Django==4.2.17
django-background-tasks==1.2.8
django-eventstream==4.5.1
django-storages==1.14.4
pillow==11.0.0
psycopg2-binary==2.9.9
opencv-python==4.10.0.84
gunicorn
whitenoise
numpy
supabase==2.7.4
nltk